finished refactoring EmModel to openmp. Hopefully there are no problems

+++ cfm-code/Solver.h
@@ -12,21 +12,21 @@
 
 class Solver {
 public:
-	virtual void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	virtual void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                           boost::shared_ptr<Param> param) = 0;
 
-	void setLearningRate(const float lr) { this->learning_rate = lr; };
+	void setLearningRate(const double lr) { this->learning_rate = lr; };
 
 	virtual ~Solver() = default;
 
 protected:
-	float learning_rate;
+	double learning_rate;
 };
 
 class Sgd : public Solver {
 public:
 	explicit Sgd(float learning_rate);
-	void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                   boost::shared_ptr<Param> param) override;
 };
 
@@ -34,7 +34,7 @@ class Momentum : public Solver {
 public:
 	Momentum(unsigned int length, float learning_rate, float momentum);
 
-	void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                   boost::shared_ptr<Param> param) override;
 
 private:
@@ -46,7 +46,7 @@ class Adam : public Solver {
 public:
 	Adam(unsigned int length, float learning_rate, float beta_1, float beta_2, float eps, bool use_amsgrad);
 
-	void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                   boost::shared_ptr<Param> param) override;
 
 protected:
@@ -69,7 +69,7 @@ public:
 		this->w = w;
 	};
 
-	void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                   boost::shared_ptr<Param> param) override;
 
 private:
@@ -83,7 +83,7 @@ public:
 	AdaBelief(unsigned int length, float learning_rate, float beta_1, float beta_2, float eps, bool use_amsgrad)
 	    : Adam(length, learning_rate, beta_1, beta_2, eps, use_amsgrad) {};
 
-	void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                   boost::shared_ptr<Param> param) override;
 };
 
@@ -91,7 +91,7 @@ class AdaDelta : public Solver {
 public:
 	AdaDelta(unsigned int length, float learning_rate, float decay_rate, float eps);
 
-	void adjustWeights(std::vector<float> &grads, std::set<unsigned int> &used_idxs,
+	void adjustWeights(std::vector<double> &grads, std::set<unsigned int> &used_idxs,
 	                   boost::shared_ptr<Param> param) override;
 
 private:
